projeto finalizado

+++ src/routes.js
@@ -9,25 +9,30 @@ export const routes = [
         method: 'GET',
         path: buildRoutePath('/tasks'),
         handler: (req, res) => {
-            const tasks = database.select('tasks')
-
-            return res.end(JSON.stringify(tasks))
+          const { search } = req.query
+    
+          const tasks = database.select('tasks', search ?{
+            title: search,
+            description: search
+          } : null)
+    
+          return res.end(JSON.stringify(tasks))
         }
-    },
+      },
 
     {
         method: 'POST',
         path: buildRoutePath('/tasks'),
         handler: (req, res) => {
-            const {title, description, completed_at, created_at, updated_at} = req.body
+            const {title, description} = req.body
 
             const task = {
                 id : randomUUID(),
                 title,
                 description,
-                completed_at,
-                created_at,
-                updated_at
+                completed_at: null,
+                created_at: new Date().toISOString(),
+                updated_at: new Date().toISOString(),
             }
 
             database.insert('tasks', task)
@@ -40,25 +45,63 @@ export const routes = [
         method: 'PUT',
         path: buildRoutePath('/tasks/:id'),
         handler: (req, res) => {
-            const { id } = req.params
-            const {title, description, completed_at, created_at, updated_at} = req.body
-
-            database.update('tasks', id, {
-                title,
-                description,
-                completed_at,
-                created_at,
-                updated_at,
-            })
-
-            return res.writeHead(204).end()
+          const { id } = req.params
+          const { title, description } = req.body
+    
+          if (!title && !description) {
+            return res.writeHead(400).end(
+              JSON.stringify({ message: 'title or description are required' })
+            )
+          }
+    
+          const [task] = database.select('tasks', { id })
+    
+          if (!task) {
+            return res.writeHead(404).end()
+          }
+    
+          database.update('tasks', id, {
+            title: title ?? task.title,
+            description: description ?? task.description,
+            updated_at: new Date()
+          })
+    
+          return res.writeHead(204).end()
         }
-    },
+      },
+      {
+        method: 'PATCH',
+        path: buildRoutePath('/tasks/:id/complete'),
+        handler: (req, res) => {
+          const { id } = req.params
+    
+          const [task] = database.select('tasks', { id })
+    
+          if (!task) {
+            return res.writeHead(404).end()
+          }
+
+          const isTaskCompleted = !!task.completed_at
+          const completed_at = isTaskCompleted ? null : new Date()
+    
+          database.update('tasks', id, {
+            completed_at
+          })
+    
+          return res.writeHead(204).end()
+        }
+      },
     {
         method: 'DELETE',
         path: buildRoutePath('/tasks/:id'),
         handler: (req, res) => {
             const { id } = req.params
+
+            const [task] = database.select('tasks', { id })
+    
+            if (!task) {
+                return res.writeHead(404).end()
+            }
 
             database.delete('tasks', id)
